dashboard

+++ app/Config/Routes.php
@@ -7,8 +7,7 @@ $routes = Services::routes();
 
 // Load the system's routing file first, so that the app and ENVIRONMENT
 // can override as needed.
-if (file_exists(SYSTEMPATH . 'Config/Routes.php'))
-{
+if (file_exists(SYSTEMPATH . 'Config/Routes.php')) {
 	require SYSTEMPATH . 'Config/Routes.php';
 }
 
@@ -36,28 +35,31 @@ $routes->setAutoRoute(false);
 //Inicio
 $routes->get('/', 'Home::index');
 //Usuarios - Login
-$routes->get('login', 'Login::index');
-$routes->get('usuario/new',             'Usuario::new');
-$routes->post('usuario',                'Usuario::create');
-$routes->get('usuario',                 'Usuario::index');
-$routes->get('usuario/(:segment)/edit', 'Usuario::edit/$1');
-$routes->put('usuario/(:segment)',      'Usuario::update/$1');
+$routes->match(['get','post'],'login', 	'Login::index', ['filter' => 'noauth']);
+$routes->get('logout',             		'Login::logout');
+$routes->get('usuario/new',             'Usuario::new', ['filter' => 'auth']);
+$routes->post('usuario',                'Usuario::create', ['filter' => 'auth']);
+$routes->get('usuario',                 'Usuario::index', ['filter' => 'auth']);
+$routes->get('usuario/(:segment)/edit', 'Usuario::edit/$1', ['filter' => 'auth']);
+$routes->put('usuario/(:segment)',      'Usuario::update/$1', ['filter' => 'auth']);
+//Dashboard
+$routes->get('dashboard',            'Dashboard::index', ['filter' => 'auth']);
 // Posts
-$routes->get('post/new',             'Post::new');
-$routes->post('post',                'Post::create');
-$routes->get('post',                 'Post::index');
+$routes->get('post/new',             'Post::new', ['filter' => 'auth']);
+$routes->post('post',                'Post::create', ['filter' => 'auth']);
+$routes->get('post',                 'Post::index', ['filter' => 'auth']);
 $routes->get('post/(:segment)',      'Post::show/$1');
-$routes->get('post/(:segment)/edit', 'Post::edit/$1');
-$routes->put('post/(:segment)',      'Post::update/$1');
-$routes->delete('post/(:segment)',   'Post::delete/$1');
+$routes->get('post/(:num)/edit', 	 'Post::edit/$1', ['filter' => 'auth']);
+$routes->put('post/(:num)',          'Post::update/$1', ['filter' => 'auth']);
+$routes->delete('post/(:num)',       'Post::delete/$1', ['filter' => 'auth']);
 //categorias
-$routes->get('categoria/new',             'Categoria::new');
-$routes->post('categoria',                'Categoria::create');
-$routes->get('categoria',                 'Categoria::index');
+$routes->get('categoria/new',             'Categoria::new', ['filter' => 'auth']);
+$routes->post('categoria',                'Categoria::create', ['filter' => 'auth']);
+$routes->get('categoria',                 'Categoria::index', ['filter' => 'auth']);
 $routes->get('categoria/(:segment)',      'Categoria::show/$1');
-$routes->get('categoria/(:segment)/edit', 'Categoria::edit/$1');
-$routes->put('categoria/(:segment)',      'Categoria::update/$1');
-$routes->delete('categoria/(:segment)',   'Categoria::delete/$1');
+$routes->get('categoria/(:segment)/edit', 'Categoria::edit/$1', ['filter' => 'auth']);
+$routes->put('categoria/(:segment)',      'Categoria::update/$1', ['filter' => 'auth']);
+$routes->delete('categoria/(:segment)',   'Categoria::delete/$1', ['filter' => 'auth']);
 
 /*
  * --------------------------------------------------------------------
@@ -72,7 +74,6 @@ $routes->delete('categoria/(:segment)',   'Categoria::delete/$1');
  * You will have access to the $routes object within that file without
  * needing to reload it.
  */
-if (file_exists(APPPATH . 'Config/' . ENVIRONMENT . '/Routes.php'))
-{
+if (file_exists(APPPATH . 'Config/' . ENVIRONMENT . '/Routes.php')) {
 	require APPPATH . 'Config/' . ENVIRONMENT . '/Routes.php';
 }
